datapath: Add support for re-entering LXC egress path after L7 LB
Allow host to tail call to lxc egress for policy enforcement after L7 LB.

This requires L7 LB proxy to set the source Endpoint ID in the socket
mark. This mark includes a new magic value that the datapath
recognizes to subject the traffic to source LXC (policy enforcement)
processing again. Note that this does not loop endlessly due to the L7
LB switching from service frontend IP to backend IP.

iptables rules are amended to avoid overwriting the new "from L7 LB"
mark with "from host" mark. L7 load balancer upstream connections may
go to the egress proxy listener of the source pod. Pod connections to
proxy listeners are already not tracked to avoid conflicts due to
reuse of the same 5-tuple on both sides of the (egress) proxy. Add
NOTRACK rules for L7 LB connections that may go to the egress proxy
for the same reason.

A new flag 'from_l7lb' is added to conntrack to enable redirecting
reply packets back to L7 LB proxy.

Add a new map for bpf_host to tail call into for LXC policy
enforcement after L7 LB.

Some kernels reached their eBPF complexity limit with the new code
compiled in. Address this by making all L7 LB related code in
bpf_lxc.c confitional on the new ENABLE_L7_LB. This allows the new
feature to be used on kernels that allow a but more complex eBPF
programs.

Checks for the need to tail call back to the source pod also in to
to-container and to-netdev programs, for traffic routed to a local or
remote pod, respectively.

Do not reuse external original source address with L7 LB. This fixes a
problem with Ingress that caused return traffic from remote-node backends
to be routed directly to the external source instead of coming back to
the Envoy listener.

Note that this causes external traffic redirected via Ingress to look
like traffic from local node to local backends, and traffic from
remote-node to remote backends. This undoubtedly complicates ingress
policy definition for the backends.

handle_policy_egress() is tail called for traffic coming from the host
proxy, so CB_IFINDEX should be zero. Assume this is the case.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>